backend/ai_service/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
anthropic[bedrock]>=0.52.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
